services/fix-suggester/requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
jinja2
diff-match-patch
psycopg2-binary
